Final Submission
This ZIP file includes the code and animated maps of the Bonus Question. I had to compress the IPYNB file as its size was more than 100MB, and GitHub would not allow me to upload it. I also have uploaded an IPYNB version of my code for your convenience, but that version does not include the animated maps widgets and only includes the written codes.