Final Submission
This IPYNB file includes the codes for all parts of the assignment but lacks the outputs for the animated maps. The related code for producing the animated maps is included but the outputs themselves have been removed from the file because including them would increase the file size to more than 100Mb, and GitHub did not allow me to upload the file. However, I have uploaded the ZIP version of the IPYNB with an animated map into the repository as well. Also, you may run the code on your machine, as all data are openly accessible. Sorry for the inconvenience and thank you for your time.